Have a preparation command
This comes about from thinking ahead as to how to handle cabal + nix.

At some point it would be nice to have the ability for this to produce
a shell.nix file, but since it typically needs/wants to be tweaked you
don't want it automatically run.

As such, the intent is:

* No build artifacts? Prepare then build.

* Have build artifacts? Just build.