Assume that all project commands are run from the project root
Not required for stack, but doesn't hurt.

In practice, we will have to do this manually.

(Not done with `commandTargets` because it doesn't have to run within
it, so we might as well just pass the directory directly.  This is
admittedly a hack from knowing how stack and cabal work, so maybe this
should change as well.)