Remove dist-checking from default
It is not _the_ build artifact to check for a particular tool.  For
example, if `cabal.sandbox.config` exists (because we've created the
sandbox) but `dist` doesn't (as we haven't built anything yet) we
would still opt-in to using sandboxes.

It might be feasible for sandboxing to _not_ be usable if `dist`
exists but `cabal.sandbox.config` doesn't (because a different Cabal
mode was used) but that's probably not worth it.