Don't get build tool versions unless explicitly required.
This is due to the overhead arising from commercialhaskell/stack#3423

Versions are still obtained for the Nix tools and GHC, because it
avoids having to change the `GlobalEnv` type for the overal
information produced, and the overhead from that is negligible.

With this change, there is only about 0.2s overhead arising from using
_jbi_, as opposed to the approximately 1.5s overhead before (whether
you were using stack or not).