Only enable benchmarks for cabal+nix when possible
I don't really like this; the behaviour should be the same no matter
what tool is used (in that benchmarks should be enabled).

Furthermore, what happens if no additional benchmarks are needed?
In that case there's no reason _not_ to enable benchmarks.

But... it's getting to be too much of a pain to quickly make a
non-benchmarking change...